feat(Socket Services): add accept offer, create offer, decline offer to socket services
This patch allow socket events to accept offer, create offer, and decline offer. The redux
action has also been created. TODOs are added to show what is suppose to happen when data
is received from socketEvent.

Note that the redux actions created do not do anything at the moment. The socket services
are setup with the correct parameters.