feat(Create Offer): feature for user to create new offer
This patch allows user to create new offer. The user needs to enter
the price and number of shares to create a new offer, a new message
pops up at the bottom.

The message is sent and received as a socket. The chatroom will be
updated according to the latest message sent.