feat(Offer): add functionality to accept offer
This patch includes the functionality for user to accept offer.
When the accept button is clicked the event is emitted to the
backend to indicate that the offer has been accepted. A socket
response is returned to indicate that the offer has successfully
been accepted. The buttons that show the offer in the conversation
will be hidden.