feat(Decline Offer): add decline offer functionality
This commit allows user to decline offer. When the other party
declines the offer the offer changes status from PENDING to REJECTED,
the message shown at the bottom is "this offer has been rejected".

The acceptOffer redux action is reused as the functionality for both
declining offer and accepting offer is the same.